Flatten the add token styles
